Evaluate the BypassGovernance permission during API authz
Previously, we were dynamically calling the internal authorization
API to check if the identity, when not an account, has the
permission to bypass the governance lock configuration. This
call sent request contexts with information we already have
when authorizing the API, except the type of identity.

This commit introduces an optimization by including the
Bypass action, when the header is set, directly when the api
is authorized. If the identity is an account, the overhead is
lower than doing another API call: no policy are evaluated.

Both PutObjectRetention and DeleteObject API calls are updated
accordingly.